requirements-macos.txt
Flask==3.1.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
mss==10.1.0
sentence-transformers==5.2.0
torch==2.9.1
rapidfuzz==3.14.3
Pillow==12.1.0
pystray==0.19.5
pyobjc==12.1
python-doctr
python-dotenv==1.0.1
sentry-sdk==2.19.2
